app/src/test/java/com/groceryreminder/AddReminderFragmentTest.java
package com.groceryreminder;

import org.junit.Before;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.robolectric.Robolectric;
import org.robolectric.RobolectricTestRunner;
import org.robolectric.annotation.Config;

import static junit.framework.Assert.assertNotNull;

@RunWith(RobolectricTestRunner.class)
@Config(emulateSdk = 18)
public class AddReminderFragmentTest {

    private MainActivity activity;

    @Before
    public void setUp() {
        this.activity = Robolectric.buildActivity(MainActivity.class).create().start().get();
    }

    @Test
    public void whenANewFragmentInstanceIsCreatedThenItIsNotNull() {
        AddReminderFragment addReminderFragment = AddReminderFragment.newInstance();

        assertNotNull(addReminderFragment);
    }

    @Test
    public void whenTheFragmentIsAttachedThenTheOnAddReminderRequestListenerShouldBeSet() {
        AddReminderFragment addReminderFragment = AddReminderFragment.newInstance();

        addReminderFragment.onAttach(activity);

        OnAddReminderRequestListener onAddReminderRequestListener = addReminderFragment.getOnAddReminderRequestListener();
        assertNotNull(onAddReminderRequestListener);
    }
}
